Use a big integer field for package file size

+++ django/thunderstore/repository/models/package_version.py
@@ -124,7 +124,7 @@ class PackageVersion(VisibilityMixin):
         upload_to=get_version_zip_filepath,
         storage=get_storage_class(settings.PACKAGE_FILE_STORAGE)(),
     )
-    file_size = models.PositiveIntegerField()
+    file_size = models.PositiveBigIntegerField()
     file_tree = models.ForeignKey(
         "storage.DataBlobGroup",
         related_name="package_versions",
